Mid-big-rewrite-of-everything commit where your author suddenly realizes that the current design won't fly after all.
The problem: the dependency system where dependency handler functions are
coupled with the `children` property of the parent works nicely when
dependencies are traversed depth-first.  But depth-first causes an
explosion of updates if we have a complex graph.

So I made a function to sort the nodes so the update is only done once.
Now I have the problem of picking which dependency handler to run for
.combine() dependencies, since I only get to run one of them for each
child and there are multiple. combine's handler doesn't even use the new
value, since it picks the values straight from the parents. This smells.

I'm going to do away with the waste and inelegance by doing something
much simpler, and perhaps also much "dirtier".

Current plan: make .newValue part of the public API/contract and let
users directly handle the internals.

Of course, users mostly just want to use map(), combine(), filter() and
friends, so they don't really care.