Fix dependency handling bug where .update()'d node, that had parents, would cause the topological sort to crash in flames.
This used to be acceptable, when that was undefined behavior to start
with. But now that we plan to kickstart generators with a single
.update(), the assumption "updated node's parent count within the update
graph is zero" is false. Therefore we need to handle cases like:

var s = stream.link(stream.ticks, stream(), function() {});

s.update();
s.tick();

Solved by not setting updated streams' parent counts to zero until
later, when all parent-child relations within the update graph have been
found.