Make test framework support test suites property. Disable console output by default.
All files that mktest gets as argument are considered test suites.

Tests have a running number that matches (by coincidence) with the
console number.

If a test breaks, it prints out the console output leading up to the
error, the error itself, and the following console output.  This way
it's hopefully easier to locate the error.